+++ src/test/utiltest/timestampHoverProvider.test.ts
@@ -57,7 +57,7 @@ describe('TimestampHoverProvider', () => {
         assert.strictEqual(result.contents.length, 1);
         assert.strictEqual(
             result.contents[0],
-            '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
+            '*Epoch Unit*: `s`  \n*UTC Timestamp*: `test-time`');
     });
 
     it('should provide correct local hover message.', async () => {
@@ -76,10 +76,32 @@ describe('TimestampHoverProvider', () => {
         assert.strictEqual(result.contents.length, 1);
         assert.strictEqual(
             result.contents[0],
-            '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
+            '*Epoch Unit*: `s`  \n*Local Timestamp*: `test-time`');
     });
 
-    it('should provide correct custom hover message.', async () => {
+    it('should provide multiple hover messages.', async () => {
+        timeConverterMock.epochToIsoLocal.returns('test-local-time');
+        timeConverterMock.epochToISOUtc.returns('test-utc-time');
+        const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
+        await config.update('targetFormat', ['local', 'utc']);
+
+        const result = await testObject.provideHover(
+            testEditor.document,
+            new vscode.Position(3, 32),
+            tokenSource.token);
+
+        assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToIsoLocal.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToISOUtc.calledOnce, true);
+        assert.strictEqual(timeConverterMock.epochToIsoLocal.firstCall.args[0], '123456789000');
+        assert.strictEqual(timeConverterMock.epochToISOUtc.firstCall.args[0], '123456789000');
+        assert.strictEqual(result.contents.length, 1);
+        assert.strictEqual(
+            result.contents[0],
+            '*Epoch Unit*: `s`  \n*Local Timestamp*: `test-local-time`  \n*UTC Timestamp*: `test-utc-time`');
+    });
+
+    it('should provide correct simple custom hover message.', async () => {
         timeConverterMock.epochToCustom.returns('test-time');
         const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
         await config.update('targetFormat', 'YYYY');
@@ -96,7 +118,71 @@ describe('TimestampHoverProvider', () => {
         assert.strictEqual(result.contents.length, 1);
         assert.strictEqual(
             result.contents[0],
-            '*Epoch Unit*: `s`  \n*Timestamp*: `test-time`');
+            '*Epoch Unit*: `s`  \n*Formatted Timestamp*: `test-time`');
+    });
+
+    describe('advanced custom hover', () => {
+
+        it('should provide correct advanced custom hover message.', async () => {
+            timeConverterMock.epochToCustom.returns('test-time');
+            const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
+            await config.update('targetFormat', [{ customFormat: 'YYYY' }]);
+
+            const result = await testObject.provideHover(
+                testEditor.document,
+                new vscode.Position(3, 32),
+                tokenSource.token);
+
+            assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[0], '123456789000');
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[1], 'YYYY');
+            assert.strictEqual(result.contents.length, 1);
+            assert.strictEqual(
+                result.contents[0],
+                '*Epoch Unit*: `s`  \n*Formatted Timestamp*: `test-time`');
+        });
+
+        it('should provide correct advanced custom hover message with name.', async () => {
+            timeConverterMock.epochToCustom.returns('test-time');
+            const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
+            await config.update('targetFormat', [{ customFormat: 'YYYY', name: 'My Name' }]);
+
+            const result = await testObject.provideHover(
+                testEditor.document,
+                new vscode.Position(3, 32),
+                tokenSource.token);
+
+            assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[0], '123456789000');
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[1], 'YYYY');
+            assert.strictEqual(result.contents.length, 1);
+            assert.strictEqual(
+                result.contents[0],
+                '*Epoch Unit*: `s`  \n*My Name*: `test-time`');
+        });
+
+        it('should provide correct advanced custom hover message with name and localization off.', async () => {
+            timeConverterMock.epochToCustom.returns('test-time');
+            const config = vscode.workspace.getConfiguration('timing.hoverTimestamp');
+            await config.update('targetFormat', [{ customFormat: 'YYYY', name: 'My Name', localize: false }]);
+
+            const result = await testObject.provideHover(
+                testEditor.document,
+                new vscode.Position(3, 32),
+                tokenSource.token);
+
+            assert.strictEqual(timeConverterMock.isValidEpoch.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.calledOnce, true);
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[0], '123456789000');
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[1], 'YYYY');
+            assert.strictEqual(timeConverterMock.epochToCustom.firstCall.args[2], false);
+            assert.strictEqual(result.contents.length, 1);
+            assert.strictEqual(
+                result.contents[0],
+                '*Epoch Unit*: `s`  \n*My Name*: `test-time`');
+        });
     });
 
     it('should return undefined if position is no epoch time.', async () => {